fixed constant loop if both 32 and 64 variants exist
Script would loop endlessly if both steam_api.dll and steam_api64.dll were found in the directory. Fixed by making it rename the original file to "steam_api[64]_o.dll" and by having it look for that file before looking to see if the other dll exists and if it does, then its sent to another section that tells you it exists and quits the script. It's also there to warn you that you've possibly already used it since the"*_o.dll" file exists already.